+++ tests/test_webtransport/main.rs
@@ -1,5 +1,6 @@
 mod capsule;
 mod flow_control;
+mod init;
 mod integration;
 mod root;
 mod stream;